Agregados castings
Por algún motivo, sin ellos no me era posible compilar

+++ src/Sources/Draw.h
@@ -14,6 +14,6 @@ protected:
 	Draw(Texture* texture=nullptr, Vector2D pos = { 0,0 }, Vector2D scale = {0,0}) :GameObject(pos, scale), texture_(texture) {};
 public:
 	virtual void draw() {};
-	virtual SDL_Rect getDestRect() { return { getPosX(), getPosY(), getScaleX(), getScaleY() }; }
+	virtual SDL_Rect getDestRect() { return { (int)getPosX(), (int)getPosY(), (int)getScaleX(), (int)getScaleY() }; }
 };
